\name{genobins}
\alias{genobins}
\docType{data}
\title{ Remove peaks from files and leave bins }
\description{
  Removes peak calls from files and leaves bin calls.  Prepares for output to statistical software.
}
\usage{genobins(samples)}
\arguments{
  \item{samples}{ a data.frame of samples containing peak calls and bin calls }
}
\details{
  Removes peak calls from files and leaves bin calls.  Prepares for output to statistical software.
}
\value{
  Returns a data.frame.
}
\references{ 
Knaus, B.J. In prep. Genomatic: an R package for DNA fragment analysis project management.
}
\author{ 
Brian J. Knaus
  \email{bknaus@fs.fed.us}
  \url{http://brianknaus.com}
}
\keyword{manip}
