\name{genogui}
\alias{genogui}
\docType{data}
\title{ Genomatic graphical user interface }
\description{
  Graphical user interface designed to run genomatic functions.
}
\usage{genogui()}
\arguments{
  \item{}{ genogui takes no arguments }
}
\details{
  Graphical user interface designed to run genomatic functions.
}
\value{
  Returns nothing.
}
\references{ 
Knaus, B.J. In prep. Genomatic: an R package for DNA fragment analysis project management.
}
\author{ 
Brian J. Knaus
  \email{bknaus@fs.fed.us}
  \url{http://brianknaus.com}
}
\keyword{manip}
