\name{genomatic2genalex}
\alias{genomatic2genalex}
\docType{data}
\title{ Converts data to GenAlEx format }
\description{
  Converts data to GenAlEx format
}
\usage{genomatic2genalex(data, outfile)}

\arguments{
  \item{data}{ a data.frame containing your data }
  \item{outfile}{ a name for your outfile }
}
\details{
 Converts data to GenAlEx format.  This software is convenient because it provides output to many other formats.  Therefore, even if you're not interested in GenAlEx you might find it to be a convenient conversion tool.  See http://www.anu.edu.au/BoZo/GenAlEx/ for information on GenAlEx.
}
\value{
  Returns a NULL.
}
\references{ 
Knaus, B.J. In prep. Genomatic: an R package for DNA fragment analysis project management.
}
\author{ 
Brian J. Knaus
  \email{bknaus@fs.fed.us}
  \url{http://brianknaus.com}
}
\keyword{manip}
