\name{genomatic2ntsys}
\alias{genomatic2ntsys}
\docType{data}
\title{ Converts data to NTSYSpc format }
\description{
  Converts data to the NTYSYSpc format
}
\usage{genomatic2ntsys(samples, outfile)}
\arguments{
  \item{samples}{ a data.frame of your data }
  \item{outfile}{ a name for your outfile }
}
\details{
 Converts data to the NTSysPC format.  See http://www.exetersoftware.com/cat/ntsyspc/ntsyspc.html for information on NTSYSpc.  NTSYSpc is commercial software.
}
\value{
  Returns a NULL.
}
\references{ 
Knaus, B.J. In prep. Genomatic: an R package for DNA fragment analysis project management.
}
\author{ 
Brian J. Knaus
  \email{bknaus@fs.fed.us}
  \url{http://brianknaus.com}
}
\keyword{manip}
