\name{order_sample}
\alias{order_sample}
\title{ Order samples alphanumerically }
\description{
  Organizes the data alpha-numerically.
}
\usage{
order_sample(data_df)
}
\arguments{
  \item{data_df}{ A data.frame of data where the first column is the sample name. }
}
\details{
A data.frame of data where the first column is the sample name.  The data.frame is reordered alpha-numerically based on sample name.
}
\value{
  A data.frame.
}
\references{ 
Knaus, B.J. In prep. Genomatic: an R package for DNA fragment analysis project management.
}
\author{ 
Brian J. Knaus
  \email{bknaus@fs.fed.us}
  \url{http://brianknaus.com}
}
\keyword{manip}
